+++ R/quadratic_plateau.R
@@ -1,5 +1,5 @@
-#' @name linear_plateau
-#' @title Linear Plateau
+#' @name quadratic_plateau
+#' @title Quadratic Plateau
 #' @description This function helps to fit a linear-plateau model in order to
 #' estimate critical soil test values (CSTV) above which yield response becomes flat.
 #' @param data Optional argument to call and object of type data.frame or data.table 
@@ -8,12 +8,12 @@
 #' @param RY name of the vector containing relative yield values (%) of type `numeric`.
 #' @param tidy logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list (default).
 #' @param resid logical operator (TRUE/FALSE) to plot residuals analysis, Default: FALSE
-#' @param plot logical operator (TRUE/FALSE) to plot the linear-plateau model, Default: FALSE
+#' @param plot logical operator (TRUE/FALSE) to plot the quadratic-plateau model, Default: FALSE
 #' @param x selfstart vector for independent variable, Default: NULL
 #' @param intercept selfstart arg. for intercept Default: NULL
 #' @param slope selfstart arg. for slope Default: NULL
 #' @param Xc selfstart arg. for critical value Default: NULL
-#' @rdname linear_plateau
+#' @rdname quadratic_plateau
 #' @return returns a `data.frame` if plot = FALSE, if plot = TRUE
 #' @details This function fits a linear-plateau model using a native selfStart function
 #' @examples 
@@ -23,7 +23,7 @@
 #'  dat <- data.frame("ry" = c(65,80,85,88,90,94,93,96,97,95,98,100,99,99,100),
 #'                    "stv" = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15))
 #'  # Run
-#'  fit_example_lp <- linear_plateau(data = dat, RY = ry, STV = stv, resid = TRUE, plot = FALSE)
+#'  fit_example_lp <- quadratic_plateau(data = dat, RY = ry, STV = stv, resid = TRUE, plot = FALSE)
 #'  fit_example_lp
 #'  }
 #' }
@@ -39,7 +39,7 @@
 #'  \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_rug}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{annotate}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}}
 #'  \code{\link[ggpp]{annotate}}
 #' @note Adapted from Austin Pearce's code. For extended reference, we recommend 
-#' to visit: https://gradcylinder.org/linear-plateau/ & https://github.com/austinwpearce/SoilTestCocaCola.
+#' to visit: https://gradcylinder.org/quad-plateau/ & https://github.com/austinwpearce/SoilTestCocaCola.
 #' @export
 #' @importFrom rlang eval_tidy quo
 #' @importFrom minpack.lm nlsLM
@@ -47,76 +47,65 @@
 #' @importFrom AICcmodavg AICc
 #' @importFrom modelr rsquare
 #' @importFrom nlstools nlsResiduals confint2
-#' @importFrom dplyr bind_cols %>%
+#' @importFrom dplyr bind_cols %>% mutate
 #' @importFrom ggplot2 ggplot aes geom_rug geom_point geom_vline geom_hline geom_path annotate scale_y_continuous labs theme_bw theme unit rel element_blank element_text
 #' @importFrom stats lm AIC optim coef predict
 #' 
 NULL
 
 
-LP_init <- function(mCall, LHS, data, ...){
+QP_init <- function(mCall, LHS, data, ...){
   
   ### STAGE 1 ====================================================================
   # Self start functions (adapted from nlraa package by Fernando Miguez)
   # Original source: https://github.com/femiguez/nlraa
-  # Define the linear-plateau function (LP_f) 
   
-  # Initialize function
+  # Initialization function
   
-  xy <- stats::sortedXyData(mCall[["x"]], LHS, data)
-  if(nrow(xy) < 4){
-    stop("Too few distinct input values to fit a linear-plateau model")
+  xy <- sortedXyData(mCall[["x"]], LHS, data)
+  if(nrow(xy) < 3){
+    stop("Too few distinct input values to fit a quadratic-platueau-3-Xc.")
   }
-  
-  # Start with a guess from simple lm to half the data
-  xy1 <- xy[1:floor(nrow(xy)/2),]
-  fit1 <- stats::lm(xy1[,"y"] ~ xy1[,"x"])
-  
-  ## Optimization of starting values
-  objfun <- function(cfs){
-    pred <- LP_f(xy[,"x"], intercept=cfs[1], slope=cfs[2], Xc=cfs[3])
-    ans <- sum((xy[,"y"] - pred)^2)
-    ans
-  }
-  # optimization
-  cfs <- c(coef(fit1),mean(xy[,"x"]))
-  op <- try(stats::optim(cfs, objfun, method = "L-BFGS-B",
-                         upper = c(Inf, Inf, max(xy[,"x"])),
-                         lower = c(-Inf, -Inf, min(xy[,"x"]))), silent = TRUE)
-  
-  if(class(op) != "try-error"){
-    intercept <- op$par[1]
-    slope <- op$par[2]
-    Xc <- op$par[3]
-  } else {
-    ## If it fails I use the mean for the CSTV (Xc)
-    ## and initial values guess by fiting a lm() to half the data
-    
-    intercept <- stats::coef(fit1)[1]
-    slope <- stats::coef(fit1)[2]
-    Xc <- mean(xy[,"x"])
-  }
-  
-  value <- c(intercept, slope, Xc)
+  ## Guess for a, b and Xc is to fit a quadratic linear regression to all the data
+  fit <- lm(xy[,"y"] ~ xy[,"x"] + I(xy[,"x"]^2))
+  a <- coef(fit)[1]
+  b <- coef(fit)[2]
+  c <- coef(fit)[3]
+  Xc <- -0.5 * b/c
+  ## If I fix a and b maybe I can try to optimze Xc only
+  value <- c(a, b, Xc)
   names(value) <- mCall[c("intercept","slope","Xc")]
   value
 }
 
-#' @rdname linear_plateau
-#' @return LP_f: vector of the same length as x using the linear-plateau function
+#' @rdname quadratic_plateau
+#' @return QP_f: vector of the same length as x using the quadratic-plateau function
 #' @export
 #' 
-LP_f <- function(x, intercept, slope, Xc){
+QP_f <- function(x, intercept, slope, Xc){
   
-  .asym <- intercept + slope * Xc
-  .value <- (x < Xc) * (intercept + slope * x) + (x >= Xc) * .asym
+  .value <- (x <= Xc) * (intercept + slope * x + (-0.5 * slope/Xc) * x^2) + (x > Xc) * (intercept + (-slope^2)/(4 * -0.5 * slope/Xc))
   
-  ## Derivative with respect to a when (x < Xc)
-  .exp1 <- 1 ## ifelse(x < Xc, 1, 1)
+  ## Derivative with respect to intercept
+  .exp1 <- 1 
+  
   ## Derivative with respect to slope
-  .exp2 <- ifelse(x < Xc, x, Xc)
+  ## .exp2 <- deriv(~ intercept + slope * x + (-0.5 * slope/Xc) * x^2, "slope")
+  ## .exp2slope <- deriv(~ intercept + (-slope^2)/(4 * -0.5 * slope/Xc), "slope")
+  .expr2 <- -slope^2
+  .expr4 <- 4 * -0.5
+  .expr6 <- .expr4 * slope/Xc
+  .exp2 <- ifelse(x <= Xc, x - 0.5/Xc * x^2, -(2 * slope/.expr6 + .expr2 * (.expr4/Xc)/.expr6^2))
+  
   ## Derivative with respect to Xc
-  .exp3 <- ifelse(x < Xc, 0, slope)
+  ## .exp2 <- deriv(~ (intercept + slope * x + (-0.5 * slope/Xc) * x^2), "Xc")
+  ## .exp2slope <- deriv(~ a + (-slope^2)/(4 * -0.5 * slope/Xc), "Xc")
+  .expr4 <- -0.5 * slope
+  .expr6 <- x^2
+  .expr2 <- -slope^2
+  .expr5 <- 4 * -0.5 * slope
+  .expr7 <- .expr5/Xc
+  .exp3 <- ifelse(x <= Xc, -(.expr4/Xc^2 * .expr6), .expr2 * (.expr5/Xc^2)/.expr7^2)
   
   .actualArgs <- as.list(match.call()[c("intercept","slope","Xc")])
   
@@ -132,15 +121,15 @@ LP_f <- function(x, intercept, slope, Xc){
   .value
 }
 
-#' @rdname linear_plateau
-#' @return SS_LP: selfStart object to pass into the linear_plateau fit
+#' @rdname quadratic_plateau
+#' @return SS_QP: selfStart object to pass into the quadratic_plateau fit
 #' @export 
-SS_LP <- stats::selfStart(LP_f, initial = LP_init, c("intercept","slope","Xc"))
+SS_QP <- stats::selfStart(QP_f, initial = QP_init, c("intercept","slope","Xc"))
 
-#' @rdname linear_plateau
-#' @return linear_plateau: function
+#' @rdname quadratic_plateau
+#' @return quadratic_plateau: function
 #' @export 
-linear_plateau <- function(data = NULL,
+quadratic_plateau <- function(data = NULL,
                                 STV,
                                 RY,
                                 tidy = FALSE,
@@ -167,49 +156,80 @@ linear_plateau <- function(data = NULL,
   miny <- min(test.data$y)
   maxy <- max(test.data$y)
   
-  # Run the model combining minpack.lm::nlsLM + defined selfStart (SS_LP)
-  lp_model <-
-    try(minpack.lm::nlsLM(y ~ SS_LP(x, b0, b1, CSTV), data = test.data))
+  # Run the model combining minpack.lm::nlsLM + defined selfStart (SS_QP)
+  qpmodel <-
+    try(minpack.lm::nlsLM(y ~ SS_QP(x, b0, b1, CSTV), data = test.data))
   
-  if (inherits(lp_model, "try-error")) {
+  if (inherits(qpmodel, "try-error")) {
     stop("Linear-plateau model did not converge. Please, consider other models.")
   } else {
-    lp_model <- lp_model
+    qpmodel <- qpmodel
   }
   
   # Find AIC and pseudo R-squared
   # AIC 
   # It makes sense because it's a sort of "simulation" (using training data) to 
   # test what would happen with out of sample data
-  AIC <- round(stats::AIC(lp_model), 0)
-  AICc <- round(AICcmodavg::AICc(lp_model), 0)
+  AIC <- round(stats::AIC(qpmodel), 0)
+  AICc <- round(AICcmodavg::AICc(qpmodel), 0)
   # R2
-  R2 <- round(modelr::rsquare(lp_model, test.data), 2)
+  R2 <- round(modelr::rsquare(qpmodel, test.data), 2)
   
   # get model coefficients
-  b0 <- stats::coef(lp_model)[[1]]
-  b1 <- stats::coef(lp_model)[[2]]
-  CSTV <- round(stats::coef(lp_model)[[3]], 1)
+  b0 <- stats::coef(qpmodel)[[1]]
+  b1 <- stats::coef(qpmodel)[[2]]
+  b2 <- -0.5 * b1 / stats::coef(qpmodel)[[3]]
+  CSTV <- round(stats::coef(qpmodel)[[3]], 1)
   # confidence interval for CSTV
-  lp_model.confint <- nlstools::confint2(lp_model)
-  CSTV_lower <- lp_model.confint[[3,1]]
-  CSTV_upper <- lp_model.confint[[3,2]]
-  plateau <- b0 + b1 * CSTV
+  qpmodel.confint <- nlstools::confint2(qpmodel)
+  CSTV_lower <- qpmodel.confint[[3,1]]
+  CSTV_upper <- qpmodel.confint[[3,2]]
+  plateau <- b0 + b1 * CSTV + b2 * (stats::coef(qpmodel)[[3]])^2
   
-  # have to make a line because the SS_LP doesn't plot right
-  lp_line <- data.frame(x = c(minx, CSTV, maxx),
-                    y = c(b0 + b1 * minx, plateau, plateau))
+  # have to make a line because the SS_QP doesn't plot right
+  qp_line <- data.frame(x = seq(minx, maxx, by = maxx/200)) %>%
+                      dplyr::mutate(y = ifelse(x < CSTV, 
+                                        b0 + b1 * x + b2*x^2,
+                                        plateau))
   
   equation <- paste0(round(b0, 1), " + ",
-                     round(b1, 2), "x if x<CSTV")
-  
+                     round(b1, 2), "x + ",
+                     round(b2, 2), "x^2 if x<CSTV")
   ## STAGE 3 ====================================================================
+  # ## CSTV at defined % of max/plateau
+  # @param target numeric value indicating yield target to find the specific CSTV
+  # # To find an X value at a given Y less than predicted plateau
+  # if (!is.null(target)) {
+  # adjust_cstv <- function(b0, b1, b2) {
+  #   CSTV <- -0.5 * b1 / b2
+  #   newplateau <-
+  #     (b0 + (b1 * CSTV) + (b2 * CSTV * CSTV)) * target / 100
+  #   newb0 <- b0 - newplateau
+  #   discriminant <- (b1 ^ 2) - (4 * newb0 * b2)
+  #   
+  #   if (discriminant < 0) {
+  #     return(NA)
+  #   }
+  #   else if (discriminant > 0) {
+  #     cstv_adj <- (-b1 + sqrt(discriminant)) / (2 * b2)
+  #     
+  #     return(cstv_adj)
+  #   }
+  # }
+  # 
+  # cstv_adj <- adjust_cstv(b0, b1, b2)
+  # 
+  # } else {cstv_adj <- NULL}
+  
+  
+  
+  ## STAGE 4 ====================================================================
   ## Outputs
   # Table output =================================================
   if (plot == FALSE) {
     {
       if (resid == TRUE)
-        plot(nlstools::nlsResiduals(lp_model), which = 0)
+        plot(nlstools::nlsResiduals(qpmodel), which = 0)
     }
     results <- data.frame(
       intercept = round(b0, 2),
@@ -224,7 +244,7 @@ linear_plateau <- function(data = NULL,
       R2
       )
     
-  # Decide type of output
+    # Decide type of output
     if (tidy == TRUE) {results <- results}
     
     if (tidy == FALSE) {results <- as.list(results)}
@@ -235,16 +255,16 @@ linear_plateau <- function(data = NULL,
     # Residual plots and normality
     {
       if (resid == TRUE)
-        plot(nlstools::nlsResiduals(lp_model), which = 0)
+        plot(nlstools::nlsResiduals(qpmodel), which = 0)
     }
     
     # Generate predicted values
-    predicted <- stats::predict(lp_model, newdata = test.data) %>%
+    predicted <- stats::predict(qpmodel, newdata = test.data) %>%
       as.data.frame() %>%
       dplyr::bind_cols(test.data)
     
     # GGPLOT output =================================================   
-    lp_plot <- predicted %>%
+    qp_plot <- predicted %>%
       ggplot2::ggplot(ggplot2::aes(x, y)) +
       ggplot2::geom_rug(alpha = 0.2, length = ggplot2::unit(2, "pt")) +
       # Data points
@@ -258,14 +278,14 @@ linear_plateau <- function(data = NULL,
       # Plateau
       ggplot2::geom_hline(yintercept = plateau, alpha = 0.2) +
       # LP Curve
-      ggplot2::geom_path(data = lp_line, ggplot2::aes(x=x,y=y), color="grey15", size = 1.5) +
+      ggplot2::geom_path(data = qp_line, ggplot2::aes(x=x,y=y), color="grey15", size = 1.5) +
       # Text annotations
       ggplot2::annotate("text",label = paste("CSTV =", CSTV, "ppm"),
                         x = CSTV, y = 0, angle = 90, hjust = 0, vjust = 1.5, col = "grey25") +
       ggplot2::annotate("text",label = paste0("Plateau = ", round(plateau, 1), "%"),
                         x = maxx, y = plateau, hjust = 1,vjust = 1.5, col = "grey25") +
       ggplot2::annotate("text", col = "grey25",
-                        label = paste0("y = ", equation, 
+                        label = paste0("y = ", equation,
                                        "\nn = ", nrow(test.data),
                                        "\npseudo-R2 = ", R2,
                                        "\nAICc = ", AICc,
@@ -273,12 +293,12 @@ linear_plateau <- function(data = NULL,
                         x = maxx, y = 0, vjust = 0, hjust = 1) +
       ggplot2::scale_y_continuous(limits = c(0, maxy),breaks=seq(0,maxy*2,10)) +
       ggplot2::labs(x = "Soil test value (units)", y = "Relative yield (%)",
-                    title = "Linear-plateau")+
+                    title = "Quadratic-plateau")+
       ggplot2::theme_bw()+
       ggplot2::theme(panel.grid = ggplot2::element_blank(),
                      axis.title = ggplot2::element_text(size = ggplot2::rel(1.5)))
     
-    return(lp_plot)
+    return(qp_plot)
   }
   
 }