+++ NEWS.md
@@ -1,7 +1,14 @@
 # soiltestR 1.0.4
 
+
+* Apr, 18, 2022. The soiltestcorr v1.0.5 includes the option for fitting selfs-starting quadratic-plateau models with the quadratic_plateau() function.
+
 * Apr, 14, 2022. The soiltestcorr v1.0.4 updated its name (previously soiltestR) in order to follow good coding practices using all lowercase letters. For the same resason, the old modALCC() function was renamed as mod_alcc().
+
 * The soiltestcorr v1.0.3 includes the option for fitting linear-plateau models, includes the 'plot' arg (boolean) in all functions, and the 'tidy' arg for choosing outputs as type list (tidy = FALSE) or data.frame (tidy = TRUE).
+
 * The soiltestcorr v1.0.2 includes the option for running the quadrants approach described by Cate & Nelson (1971).
+
 * The soiltestcorr v1.0.1 includes the option for running the quadrants approach described by Cate & Nelson (1965).
+
 * The soiltestcorr v1.0.0 includes the option for running the modified arcsine-log calibration curve described Correndo et al. (2017).